modules/requirements.txt
streamlit>=1.32.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.3
PyPDF2>=3.0.1
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers>=2.5.0   # free local embeddings — always used
anthropic>=0.18.0               # primary LLM provider (recommended)
openai>=1.12.0                  # optional fallback LLM (NOT used for embeddings)
python-dotenv>=1.0.0
